// apps/web/src/api/__generated__/types/PromotionResponseListItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * sync
 * OpenAPI spec version: 0.0.1
 */
import type { PromotionResponseListItemFieldsItem } from './PromotionResponseListItemFieldsItem';

export type PromotionResponseListItem = {
  /** 생성 시각 */
  createdAt: string;
  /**
   * 연결된 게시글의 슬러그
   * @nullable
   */
  postSlug?: string | null;
  /** 활성화 여부 */
  active: boolean;
  /**
   * 연결된 게시글의 제목
   * @nullable
   */
  postTitle?: string | null;
  /** 프로모션 ID */
  id: number;
  /** 신청 시 수집하는 필드 목록 */
  fields: PromotionResponseListItemFieldsItem[];
  /**
   * 연결된 게시글이 속한 프로젝트 핸들
   * @nullable
   */
  projectHandle?: string | null;
};